// src/types/enums/defaultModules.ts
export enum DefaultUserModules {
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  OPENPOSITION = 'OPENPOSITION',
  MENTORING = 'MENTORING',
  BENEFIT = 'BENEFIT',
  CONFIGURATION = 'CONFIGURATION',
  LEARNINGPATHS = 'LEARNINGPATHS',
  LEADS = 'LEADS',
  HANDBOOK = 'HANDBOOK',
  TECH_INTERVIEW_RECRUITER = 'TECHINTERVIEW/RECRUITER',
  TECH_INTERVIEW_STACK_MANAGER = 'TECHINTERVIEW/STACKMANAGER',
  TECH_INTERVIEW_INTERVIEWER = 'TECHINTERVIEW/INTERVIEWER',
}